Modified Makefile to create a shared library
Modified Makefile to be able to create a shared library. This simplifies
the use of that code (used in the example) in other projects. Now, the
example provided links against this library as well.

Because Raspberry Pi and Jetson differ in hardware, they use different
libraries, specific for each platform. So, I have added a simple rule
which tells the user to make for one or for the other, instead of making
for both.

I have added a gitignore that ignores de bin folder, which is not
necessary to be under source control. The two binaries, epd (example)
and the library are ignored as well. I have modified the Makefile
accordingly to be able to create the bin folder.

Finally, I have added a little print in the examples' main.c to tell the
user to uncomment one of the tests, because by default, the example demo
does nothing and do not inform the user.