fix_python_cleanup_exit: Fix Python cleanup and exit execution path for some examples.
This fix apparently seems to be very trivial but if the user terminates
an executing Python example program with CTRL+C and then leaves the display
at that state then power is still applied to the display as a result of not
executing the cleanup and exit path. This can damage the display eventually.