Fix off by 1 error in Displaypart
There was an off by one error requiring all pattern buffers to have an additional byte at the start. The error has been updated and the function has been documented for other users.

Also added the inversion "~" of the white bytes to keep with the same format in the pattern buffer.