Fix PWR issues when booting up on RPI
I noticed quite often, that a RPI is dying multiple times (red LED), whenever it hits this code. It seems, that due to the fact that the PWR PIN is up, and while trying to set default values for the other PINs, this creates a lot of issues. To the extend that the RPI turns off. Thus, first turn off PWR before setting GPIO default values.